Tests/LibWeb/Text/expected/WebAssembly-Memory-grow-shared-stale-view.txt

firstView[0x0]: 0x11
firstView[0x1]: 0x22
firstView[0x2]: 0x33
firstView[0x3]: 0x44
thirdView[0x4]: 0x55
thirdView[0x5]: 0x66
thirdView[0x6]: 0x77
thirdView[0x7]: 0x88
firstView[0x8]: 0x99
firstView[0x9]: 0xaa
firstView[0xa]: 0xbb
firstView[0xb]: 0xcc
